loki_config.c Enhancements/fixes:
- COMMON_OPTIONS would have incorrect value if WINDOWED_ONLY was defined.
Fixed.

- MAX_OPTIONS has been increased from 16 to 255. I doubt it's good design
to use that many command line options (*coughcoughDESCENT3cough), but I
actually ran out with 16 (note that this number includes the 4-6 "common"
options.)

- Removed an inconsistency: even though the common options are stored in
the
array with the application-registered options, and we make allowances for
their existance, we don't use them in the usage output, and opt for
hardcoded
strings instead. I changed it to be the same as registering options,
although
they are their automagically at startup. This greatly streamlines the
whole
process.

- Rewrote the usage output code. The use of '\t' characters didn't align
option comments correctly if your option's name was too small ("--cdrom",
for example), so it actually calculates the correct amount of ' ' chars to
output.

- Other odd condition fixes: output is more correct in cases where
specific
command line option elements are omitted.

--ryan.